chore: update some dep & fix some error
