+++ nft-glue/src/index.ts
@@ -2,28 +2,26 @@ import { env } from "process";
 
 import { ecdsaSign } from "secp256k1";
 import { Reader, RPC, normalizers } from "ckb-js-toolkit";
+import { commons, Indexer } from "@ckb-lumos/lumos";
 import {
   HexString,
   Hash,
   Address,
   Cell,
   CellDep,
-  OutPoint,
   Script,
   core,
   utils,
 } from "@ckb-lumos/base";
 const { CKBHasher } = utils;
-import { common } from "@ckb-lumos/common-scripts";
 import { getConfig, initializeConfig } from "@ckb-lumos/config-manager";
 import {
   parseAddress,
-  generateAddress,
   sealTransaction,
   TransactionSkeletonType,
   TransactionSkeleton,
+  encodeToAddress,
 } from "@ckb-lumos/helpers";
-import { Indexer } from "@ckb-lumos/indexer";
 
 const CKB_RPC = "http://127.0.0.1:8114";
 
@@ -115,16 +113,18 @@ export async function generateNftToken(
   // can be used to mark certain fields as fixed, meaning no further optimizations should
   // alter those components.
   skeleton = skeleton.update("fixedEntries", (fixedEntries) => {
-    return fixedEntries.push(
-      {
-        field: "outputs",
-        index: 0,
-      }
-    );
+    return fixedEntries.push({
+      field: "outputs",
+      index: 0,
+    });
   });
   // Now let's inject input cells to the transaction so as to provide capacities
   // needed by the newly created input cells.
-  skeleton = await common.injectCapacity(skeleton, 0, [fromAddress]);
+  skeleton = await commons.secp256k1Blake160.injectCapacity(
+    skeleton,
+    0,
+    fromAddress
+  );
   // Now we can generate and fill in the correct NFT token ID.
   const hasher = new CKBHasher();
   let inputCell = skeleton.get("inputs")!.get(0)!;
@@ -146,21 +146,21 @@ export async function generateNftToken(
   });
   // The first input must be fixed as well, since it is used to generate NFT ID
   skeleton = skeleton.update("fixedEntries", (fixedEntries) => {
-    return fixedEntries.push(
-      {
-        field: "inputs",
-        index: 0,
-      }
-    );
+    return fixedEntries.push({
+      field: "inputs",
+      index: 0,
+    });
   });
   // Since we are using the NFT script, we need to include NFT cell dep.
   skeleton = skeleton.update("cellDeps", (cellDeps) => {
     return cellDeps.push(buildNftCellDep());
   });
   // Similar to injectCapacity, lumos also provides helper methods to inject fee:
-  skeleton = await common.payFee(skeleton, [fromAddress], FEE);
+  skeleton = await commons.secp256k1Blake160.payFee(skeleton, fromAddress, FEE);
   // Finally, let's generate messages that are required in transaction signing phase:
-  skeleton = common.prepareSigningEntries(skeleton, { config: CONFIG });
+  skeleton = commons.secp256k1Blake160.prepareSigningEntries(skeleton, {
+    config: CONFIG,
+  });
   return skeleton;
 }
 
@@ -253,12 +253,14 @@ export async function transferNftToken(
     return cellDeps.push(buildNftCellDep());
   });
   // For simplicity, token sender would pay for transaction fee.
-  skeleton = await common.payFee(
+  skeleton = await commons.secp256k1Blake160.payFee(
     skeleton,
-    [generateAddress(nftCell.cell_output.lock)],
+    encodeToAddress(nftCell.cell_output.lock),
     FEE
   );
   // Finally, let's generate messages that are required in transaction signing phase:
-  skeleton = common.prepareSigningEntries(skeleton, { config: CONFIG });
+  skeleton = commons.secp256k1Blake160.prepareSigningEntries(skeleton, {
+    config: CONFIG,
+  });
   return skeleton;
 }
