+++ nft-validator/Cargo.toml
@@ -1,7 +1,9 @@
 [workspace]
-members = [
-    "tests",
-]
-exclude = [
-    "contracts",
-]
+members = ["tests", "contracts/nft-validator"]
+
+[profile.release]
+overflow-checks = true
+opt-level = 's'
+lto = true
+codegen-units = 1
+panic = 'abort'
