+++ nft-validator/capsule.toml
@@ -1,3 +1,13 @@
+# [rust]
+# # path of rust contracts workspace directory,
+# # a `Cargo.toml` file is expected under the directory.
+# workspace_dir = "."
+# toolchain = "nightly-2021-12-25"
+# docker_image = "thewawar/ckb-capsule:2021-12-25"
+
+# capsule version
+version = "0.7.3"
+# path of deployment config file
 deployment = "deployment.toml"
 
 [[contracts]]
