+++ nft-validator/contracts/nft-validator/Cargo.toml
@@ -6,12 +6,6 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-blake2b-ref = "0.1.0"
-ckb-std = "0.4.1"
+ckb-std = "0.9.0"
+blake2b-ref = "0.3.0"
 
-[profile.release]
-overflow-checks = true
-opt-level = 's'
-lto = true
-codegen-units = 1
-panic = 'abort'
