+++ nft-validator/tests/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-blake2b-ref = "0.1.0"
-ckb-tool = { git = "https://github.com/jjyr/ckb-tool.git", tag = "v0.0.2" }
-ckb-testtool = { git = "https://github.com/jjyr/ckb-tool.git", tag = "v0.0.2" }
+blake2b-ref = "0.3.0"
+ckb-tool = "0.3"
+ckb-testtool = "0.3"
 rand = "0.7.3"
